feat: now there is a styled navbar and the admin panel now also has styling
A navbar was added in base.html

functions index and tours in views.py were modified to return index_selector and
tours_selector varaibles with the template respectively. They are used
to highlight selected pages in the navbar.

File admin_tours was modified to add style to the admin panel.